refactor: extract STIG removal logic and file drop handling into composables
